Deduplicate some Linux code common to x86 and ARM
It's always better to split later if needed than to maintain exact
duplicates.
