+++ plugins/DebuggerCore/unix/linux/arch/x86-generic/PlatformThread.cpp
@@ -32,7 +32,6 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #include <elf.h>
 #include <linux/uio.h>
 #include <sys/ptrace.h>
-#include <sys/syscall.h>   /* For SYS_xxx definitions */
 #include <sys/user.h>
 
 
@@ -160,160 +159,6 @@ bool PlatformThread::fillStateFromSimpleRegs(PlatformState* state) {
 		perror("PTRACE_GETREGS failed");
 		return false;
 	}
-}
-
-
-//------------------------------------------------------------------------------
-// Name: PlatformThread
-// Desc:
-//------------------------------------------------------------------------------
-PlatformThread::PlatformThread(DebuggerCore *core, IProcess *process, edb::tid_t tid) : core_(core), process_(process), tid_(tid) {
-	assert(process);
-	assert(core);
-}
-
-//------------------------------------------------------------------------------
-// Name:
-// Desc:
-//------------------------------------------------------------------------------
-PlatformThread::~PlatformThread() {
-}
-
-//------------------------------------------------------------------------------
-// Name:
-// Desc:
-//------------------------------------------------------------------------------
-edb::tid_t PlatformThread::tid() const {
-	return tid_;
-}
-
-//------------------------------------------------------------------------------
-// Name:
-// Desc:
-//------------------------------------------------------------------------------
-QString PlatformThread::name() const  {
-	struct user_stat thread_stat;
-	int n = get_user_stat(QString("/proc/%1/task/%2/stat").arg(process_->pid()).arg(tid_), &thread_stat);
-	if(n >= 2) {
-		return thread_stat.comm;
-	}
-
-	return QString();
-}
-
-//------------------------------------------------------------------------------
-// Name:
-// Desc:
-//------------------------------------------------------------------------------
-int PlatformThread::priority() const  {
-	struct user_stat thread_stat;
-	int n = get_user_stat(QString("/proc/%1/task/%2/stat").arg(process_->pid()).arg(tid_), &thread_stat);
-	if(n >= 30) {
-		return thread_stat.priority;
-	}
-
-	return 0;
-}
-
-//------------------------------------------------------------------------------
-// Name:
-// Desc:
-//------------------------------------------------------------------------------
-edb::address_t PlatformThread::instruction_pointer() const  {
-	struct user_stat thread_stat;
-	int n = get_user_stat(QString("/proc/%1/task/%2/stat").arg(process_->pid()).arg(tid_), &thread_stat);
-	if(n >= 18) {
-		return thread_stat.kstkeip;
-	}
-
-	return 0;
-}
-
-//------------------------------------------------------------------------------
-// Name:
-// Desc:
-//------------------------------------------------------------------------------
-QString PlatformThread::runState() const  {
-	struct user_stat thread_stat;
-	int n = get_user_stat(QString("/proc/%1/task/%2/stat").arg(process_->pid()).arg(tid_), &thread_stat);
-	if(n >= 3) {
-		switch(thread_stat.state) {           // 03
-		case 'R':
-			return tr("%1 (Running)").arg(thread_stat.state);
-			break;
-		case 'S':
-			return tr("%1 (Sleeping)").arg(thread_stat.state);
-			break;
-		case 'D':
-			return tr("%1 (Disk Sleep)").arg(thread_stat.state);
-			break;
-		case 'T':
-			return tr("%1 (Stopped)").arg(thread_stat.state);
-			break;
-		case 't':
-			return tr("%1 (Tracing Stop)").arg(thread_stat.state);
-			break;
-		case 'Z':
-			return tr("%1 (Zombie)").arg(thread_stat.state);
-			break;
-		case 'X':
-		case 'x':
-			return tr("%1 (Dead)").arg(thread_stat.state);
-			break;
-		case 'W':
-			return tr("%1 (Waking/Paging)").arg(thread_stat.state);
-			break;
-		case 'K':
-			return tr("%1 (Wakekill)").arg(thread_stat.state);
-			break;
-		case 'P':
-			return tr("%1 (Parked)").arg(thread_stat.state);
-			break;
-		default:
-			return tr("%1").arg(thread_stat.state);
-			break;
-		}
-	}
-
-	return tr("Unknown");
-}
-
-//------------------------------------------------------------------------------
-// Name: step
-// Desc: steps this thread one instruction, passing the signal that stopped it
-//       (unless the signal was SIGSTOP)
-//------------------------------------------------------------------------------
-void PlatformThread::step() {
-	core_->ptrace_step(tid_, resume_code(status_));
-}
-
-//------------------------------------------------------------------------------
-// Name: step
-// Desc: steps this thread one instruction, passing the signal that stopped it
-//       (unless the signal was SIGSTOP, or the passed status != DEBUG_EXCEPTION_NOT_HANDLED)
-//------------------------------------------------------------------------------
-void PlatformThread::step(edb::EVENT_STATUS status) {
-	const int code = (status == edb::DEBUG_EXCEPTION_NOT_HANDLED) ? resume_code(status_) : 0;
-	core_->ptrace_step(tid_, code);
-}
-
-//------------------------------------------------------------------------------
-// Name: resume
-// Desc: resumes this thread, passing the signal that stopped it
-//       (unless the signal was SIGSTOP)
-//------------------------------------------------------------------------------
-void PlatformThread::resume() {
-	core_->ptrace_continue(tid_, resume_code(status_));
-}
-
-//------------------------------------------------------------------------------
-// Name: resume
-// Desc: resumes this thread , passing the signal that stopped it
-//       (unless the signal was SIGSTOP, or the passed status != DEBUG_EXCEPTION_NOT_HANDLED)
-//------------------------------------------------------------------------------
-void PlatformThread::resume(edb::EVENT_STATUS status) {
-	const int code = (status == edb::DEBUG_EXCEPTION_NOT_HANDLED) ? resume_code(status_) : 0;
-	core_->ptrace_continue(tid_, code);
 }
 
 //------------------------------------------------------------------------------
@@ -469,24 +314,5 @@ unsigned long PlatformThread::get_debug_register(std::size_t n) {
 long PlatformThread::set_debug_register(std::size_t n, long value) {
 	return ptrace(PTRACE_POKEUSER, tid_, offsetof(struct user, u_debugreg[n]), value);
 }
-
-//------------------------------------------------------------------------------
-// Name: stop
-// Desc:
-//------------------------------------------------------------------------------
-void PlatformThread::stop() {
-	syscall(SYS_tgkill, process_->pid(), tid_, SIGSTOP);
-	// TODO(eteran): should this just be this?
-	//::kill(tid_, SIGSTOP);
-}
-
-//------------------------------------------------------------------------------
-// Name: isPaused
-// Desc: returns true if this thread is currently in the debugger's wait list
-//------------------------------------------------------------------------------
-bool PlatformThread::isPaused() const {
-	return core_->waited_threads_.contains(tid_);
-}
-
 
 }
